Refactor de AuthPlayer et StockagePlayerDatabase et retrait de ces fichiers
