+++ src/main/java/games/project/koala_rock/View/VueConnexion.java
@@ -1,6 +1,6 @@
 package games.project.koala_rock.View;
 
-import games.project.koala_rock.Metier.entite.AuthPlayer;
+import games.project.metier.entite.AuthPlayer;
 import games.project.koala_rock.Metier.manager.PlayerManager;
 import games.project.koala_rock.Model.Fond;
 import games.project.koala_rock.RessourcesAccess;
